requirements.txt
# Core dependencies
Flask>=2.2.0
tweepy>=4.14.0
textblob>=0.17.1
matplotlib>=3.7.0
tqdm>=4.65.0
nltk>=3.8.1
python-dotenv>=1.0.0

# Optional: for advanced analysis or better models
# vaderSentiment>=3.3.2
# transformers>=4.40.0
# torch>=2.0.0
